refactor the use of add task reducer
